Configurando Cache

+++ app/Http/Controllers/ProductController.php
@@ -4,12 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ProductRequest as Request;
 use App\Models\Product;
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Cache;
 
 class ProductController extends Controller
 {
     public function index()
     {
-        return Product::all();
+        $minutes = Carbon::now()->addMinutes(10);
+
+        $products = Cache::remember('api:products', $minutes, function(){
+            return Product::all();
+        });
+
+        return $products;
     }
 
     public function show(Product $product)
@@ -19,11 +28,18 @@ class ProductController extends Controller
 
     public function store(Request $request)
     {
-        return Product::create($request->all());
+        Cache::forget('api:products');
+        
+        $data = $request->all();
+        $data['user_id'] = Auth::user()->id;
+
+        return Product::create($data);
     }
 
     public function update(Request $request, Product $product)
     {
+        Cache::forget('api:products');
+
         if ($product->update($request->all()) ){
             return $product;
         }
@@ -35,6 +51,9 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+        Cache::forget('api:products');
+
+        $this->authorize('delete', $product);
         $product->delete();
     }
 }
